Refactoring: ranges extracted as chess-independent
