+++ src/main/java/com/leokom/games/chess/player/legal/brain/normalized/ProtectionEvaluator.java
@@ -5,6 +5,7 @@ import com.leokom.games.chess.engine.Position;
 import com.leokom.games.chess.player.legal.brain.common.Evaluator;
 import com.leokom.games.chess.player.legal.brain.common.EvaluatorType;
 import com.leokom.games.chess.player.legal.brain.denormalized.DenormalizedEvaluatorFactory;
+import com.leokom.games.commons.brain.normalized.range.SymmetricalNormalizedRange;
 
 /**
  * Author: Leonid
